+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # 1.x branch
+## 1.5 branch
+### 1.5.0
+* fixed bug, added template for ajax requests;
+* fixed bugs to not rewrite existing routes;
+* added the `InstagramTrait`, inheriting methods from the `Instagram` utility
+    (which continues to exist, using the trait itself);
+* added the `InstagramComponent`;
+* the `PhotosWidgetsCell` class uses the `__construct()` method to set an
+    `Instagram` instance;
+* updated for MeCms 2.18.2.
+
 ## 1.4 branch
 ### 1.4.1
 * updated for CakePHP 3.4.